LPS-106243 Move SEO logic into SEO display context
